改了倒地自起会跑死的bug，上板让倒地自起时不会疯，下一步是加掉线保护

+++ others/Motion_Control/Ctrl/inc/Leg_Control.h
@@ -5,7 +5,7 @@
 #include "VMC.h"
 #include <stdint.h>
 
-extern int HZ;
+extern uint16_t HZ;
 extern float torque_limit;
 extern float torque_release_rate;
 extern float stuck_counter_time_s;
